+++ pages/index.js
@@ -1,34 +1,9 @@
-import { Box, Button, Text, TextField, Image } from '@skynexui/components';
-import appConfig from '../config.json';
+import { Box, Button, Text, TextField, Image } from '@skynexui/components'
+import React from 'react'
+import {useRouter} from 'next/router'
+import appConfig from '../config.json'
 
-function GlobalStyle() {
-  return (
-    <style global jsx>{`
-      * {
-        margin: 0;
-        padding: 0;
-        box-sizing: border-box;
-        list-style: none;
-      }
-      body {
-        font-family: 'Open Sans', sans-serif;
-      }
-      /* App fit Height */ 
-      html, body, #__next {
-        min-height: 100vh;
-        display: flex;
-        flex: 1;
-      }
-      #__next {
-        flex: 1;
-      }
-      #__next > * {
-        flex: 1;
-      }
-      /* ./App fit Height */ 
-    `}</style>
-  );
-}
+
 
 function Titulo(props) {
   const Tag = props.tag || 'h1';
@@ -60,11 +35,12 @@ function Titulo(props) {
 // export default HomePage
 
 export default function PaginaInicial() {
-  const username = 'kishica';
+  //const username = 'kishica';
+  const [username, setUsername] = React.useState('kishica')
+  const router = useRouter()
 
   return (
     <>
-      <GlobalStyle />
       <Box
         styleSheet={{
           display: 'flex', alignItems: 'center', justifyContent: 'center',
@@ -91,6 +67,11 @@ export default function PaginaInicial() {
           {/* Formulário */}
           <Box
             as="form"
+            onSubmit={function(event){
+              event.preventDefault()
+              console.log('Alguém submeteu o form')
+              router.push('/chat')
+            }}
             styleSheet={{
               display: 'flex', flexDirection: 'column', alignItems: 'center', justifyContent: 'center',
               width: { xs: '100%', sm: '50%' }, textAlign: 'center', marginBottom: '32px',
@@ -101,7 +82,22 @@ export default function PaginaInicial() {
               {appConfig.name}
             </Text>
 
+            {/* <input 
+              type="text"
+              value={username}
+              onChange={function Handler(event){
+                console.log('Usuário digitou', event.target.value)
+                const valor = event.target.value
+                setUsername(valor)
+              }}
+            /> */}
             <TextField
+              value={username}
+              onChange={function Handler(event){
+                console.log('Usuário digitou', event.target.value)
+                const valor = event.target.value                
+                setUsername(valor)
+              }}            
               fullWidth
               textFieldColors={{
                 neutral: {
